Proof of concept for ViewModel\Converter
Useful for Converters that work on a whole Model. Samples are extracted from api-raml and deserialized into Model objects, then converted to every available View Model.
We then test that the View Model is coherent with what the Converter supports(), and that it can be serialized to an array. We may find stricter properties about it