Run all end2end tests
And not only the subset tagged with 'two'. Now that journal and elife 2.0 are integrated with the 1.0 projects, the end2end tests should be the same.

At the moment, the end2end tests try to load an HTML page on journal for each article being published (not checking yet for images and other media)