Coupling to the ORCID OAuth response
Currently profiles is exposing everything that comes through the ORCID OAuth token response. `orcid` and `name` are not part of the standard, so it's not mandatory for our `profile` implementation to provide them. And even if it wants to provide these values as part of the token response, they don't necessarily need to be named/modelled like the ORCID fields.