Upgrade broken node jessie image
The current `node` image is based on Debian Jessie, which has gone into LTS only support and removed the `jessie-updates` channel (is that the right word). Any `apt-get update` within this image fails with 404.

At the time of writing `node:6.17.0-jessie` is broken too, but it is the
latest . So we switch to the latest Node.js 6.x available on the next
(and currently only `stable`) [Debian release](https://wiki.debian.org/DebianReleases), Stretch.