feat: update EventContentHeaderConverter to simplify date creation (#1825)
* feat: update EventContentHeaderConverter to simplify date creation

Removed the use of 'CreatesDate' in 'EventContentHeaderConverter'. The
'meta' variable is now directly assigned with the object's start date
via 'ViewModel\Date::simple' instead of using a conditional statement.

* feat: update event date in EventControllerTest

Update test data in EventControllerTest to use a future date. This change
updates the start date of the event to 'Sep 6, 2023' and modifies the
duration of the event to last for one day.