Hard code the path to the preload.json file in the Symfony DependencyInjection
There may a better way to obtain this path than hard coding it, it seems brittle, but it works for now. The current aim is to get the app returning status 200 pages in a browser. More robust derivation of this path can be addressed later, if we get that far.